third_party/WebKit/LayoutTests/inspector/console/console-filter-level-test-expected.txt

CONSOLE MESSAGE: line 10: sample info
CONSOLE MESSAGE: line 11: sample log
CONSOLE WARNING: line 12: sample warning
CONSOLE DEBUG: line 13: sample debug
CONSOLE ERROR: line 14: sample error
CONSOLE MESSAGE: line 16: abc info
CONSOLE MESSAGE: line 17: def info
CONSOLE WARNING: line 19: abc warn
CONSOLE WARNING: line 20: def warn
Tests that console can filter messages by source.


Running: dumpLevels
All levels
{
    error : true
    info : true
    verbose : true
    warning : true
}
Default levels
{
    error : true
    info : true
    verbose : false
    warning : true
}

Running: beforeFilter
beforeFilter
Level menu: Default levels
>console-filter-level-test.html:5 sample info
>console-filter-level-test.html:6 sample log
>console-filter-level-test.html:7 sample warning
onload @ console-filter-level-test.html:7
>console-filter-level-test.html:9 sample error
onload @ console-filter-level-test.html:9
>console-filter-level-test.html:11 abc info
>console-filter-level-test.html:12 def info
>console-filter-level-test.html:14 abc warn
onload @ console-filter-level-test.html:14
>console-filter-level-test.html:15 def warn
onload @ console-filter-level-test.html:15
>'Should be always visible'
>"Should be always visible"

Running: allLevels
Level menu: All levels
>console-filter-level-test.html:5 sample info
>console-filter-level-test.html:6 sample log
>console-filter-level-test.html:7 sample warning
onload @ console-filter-level-test.html:7
>console-filter-level-test.html:8 sample debug
>console-filter-level-test.html:9 sample error
onload @ console-filter-level-test.html:9
>console-filter-level-test.html:11 abc info
>console-filter-level-test.html:12 def info
>console-filter-level-test.html:14 abc warn
onload @ console-filter-level-test.html:14
>console-filter-level-test.html:15 def warn
onload @ console-filter-level-test.html:15
>'Should be always visible'
>"Should be always visible"

Running: defaultLevels
Level menu: Default levels
>console-filter-level-test.html:5 sample info
>console-filter-level-test.html:6 sample log
>console-filter-level-test.html:7 sample warning
onload @ console-filter-level-test.html:7
>console-filter-level-test.html:9 sample error
onload @ console-filter-level-test.html:9
>console-filter-level-test.html:11 abc info
>console-filter-level-test.html:12 def info
>console-filter-level-test.html:14 abc warn
onload @ console-filter-level-test.html:14
>console-filter-level-test.html:15 def warn
onload @ console-filter-level-test.html:15
>'Should be always visible'
>"Should be always visible"

Running: verbose
Level menu: Verbose only
>console-filter-level-test.html:8 sample debug
>'Should be always visible'
>"Should be always visible"

Running: info
Level menu: Info only
>console-filter-level-test.html:5 sample info
>console-filter-level-test.html:6 sample log
>console-filter-level-test.html:11 abc info
>console-filter-level-test.html:12 def info
>'Should be always visible'
>"Should be always visible"

Running: warningsAndErrors
Level menu: Custom levels
>console-filter-level-test.html:7 sample warning
onload @ console-filter-level-test.html:7
>console-filter-level-test.html:9 sample error
onload @ console-filter-level-test.html:9
>console-filter-level-test.html:14 abc warn
onload @ console-filter-level-test.html:14
>console-filter-level-test.html:15 def warn
onload @ console-filter-level-test.html:15
>'Should be always visible'
>"Should be always visible"

Running: abcMessagePlain
Level menu: Verbose only
>'Should be always visible'
>"Should be always visible"

Running: abcMessageRegex
Level menu: Verbose only
>'Should be always visible'
>"Should be always visible"

Running: abcMessageRegexWarning
Level menu: Warnings only
>console-filter-level-test.html:14 abc warn
onload @ console-filter-level-test.html:14
>'Should be always visible'
>"Should be always visible"
